# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VeriCell API",
    "description": "Proof of authorship, integrity and time for any digital project, anchored on Nervos CKB. Public read-only endpoints — see TECHNICAL.md §7.1.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/webhooks": {
      "post": {
        "summary": "Register a webhook for committed/consumed/superseded events",
        "tags": [
          "webhooks"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "events": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "committed",
                        "consumed",
                        "superseded"
                      ]
                    },
                    "minItems": 1
                  },
                  "unid": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "url",
                  "events"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/webhooks/{id}": {
      "delete": {
        "summary": "Remove a webhook (only the API key that registered it may delete it)",
        "tags": [
          "webhooks"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "VeriCell API (testnet)"
    }
  ],
...
}